requirements.txt
aiofiles==23.2.1
aiohttp==3.9.5
aiosignal==1.3.1
altair==5.3.0
annotated-types==0.6.0
anyio==4.3.0
arxiv==2.1.0
asgiref==3.8.1
asttokens==2.4.1
attrs==23.2.0
backoff==2.2.1
bcrypt==4.1.3
beautifulsoup4==4.12.3
blinker==1.8.1
blis==1.0.1
boltons==23.1.1
Brotli==1.1.0
build==1.2.2.post1
cachetools==5.3.3
catalogue==2.0.10
certifi==2024.2.2
cffi==1.16.0
chardet==5.2.0
charset-normalizer==3.3.2
chroma-hnswlib==0.7.6
chromadb==0.6.3
click==8.1.7
cloudpathlib==0.20.0
colorama==0.4.6
coloredlogs==15.0.1
comm==0.2.1
confection==0.1.5
contourpy==1.2.0
cssselect==1.2.0
cssselect2==0.7.0
curl_cffi==0.6.3
cycler==0.12.1
cymem==2.0.8
dataclasses-json==0.6.6
debugpy==1.8.0
decorator==5.1.1
deepdiff==7.0.1
Deprecated==1.2.18
deprecation==2.1.0
distro==1.9.0
dnspython==2.6.1
docopt==0.6.2
docutils==0.20.1
duckduckgo_search==5.3.0
durationpy==0.9
email_validator==2.1.1
emoji==2.12.1
environs==11.0.0
executing==2.0.1
extra-streamlit-components==0.1.71
faiss-cpu==1.10.0
fastapi==0.111.0
fastapi-cli==0.0.4
feedfinder2==0.0.4
feedparser==6.0.10
filelock==3.14.0
filetype==1.2.0
flatbuffers==25.2.10
fonttools==4.49.0
frozenlist==1.4.1
fsspec==2024.5.0
gitdb==4.0.11
GitPython==3.1.43
google-ai-generativelanguage==0.6.4
google-api-core==2.19.0
google-api-python-client==2.130.0
google-auth==2.29.0
google-auth-httplib2==0.2.0
google-generativeai==0.5.4
googleapis-common-protos==1.63.0
gotrue==2.4.2
graphviz==0.20.3
greenlet==3.0.3
grpcio==1.64.0
grpcio-status==1.62.2
h11==0.14.0
html5lib==1.1
htmldocx==0.0.6
httpcore==1.0.5
httplib2==0.22.0
httptools==0.6.1
httpx==0.27.0
huggingface-hub==0.23.2
humanfriendly==10.0
idna==3.6
importlib_metadata==8.4.0
importlib_resources==6.5.2
iniconfig==2.0.0
ipykernel==6.28.0
ipython==8.20.0
jedi==0.19.1
jieba3k==0.35.1
Jinja2==3.1.4
jiter==0.5.0
joblib==1.4.2
jsonlines==4.0.0
jsonpatch==1.33
jsonpath-python==1.0.6
jsonpointer==2.4
jsonschema==4.22.0
jsonschema-specifications==2023.12.1
jupyter_client==8.6.0
jupyter_core==5.7.1
Kivy==2.3.0
kivy-deps.angle==0.4.0
kivy-deps.glew==0.3.1
kivy-deps.sdl2==0.7.0
Kivy-Garden==0.1.5
kiwisolver==1.4.5
kubernetes==32.0.1
langchain==0.2.1
langchain-community==0.2.1
langchain-core==0.2.1
langchain-google-genai==1.0.5
langchain-openai==0.1.7
langchain-text-splitters==0.2.0
langcodes==3.4.1
langdetect==1.0.9
langsmith==0.1.63
language_data==1.2.0
lxml==5.2.2
lxml_html_clean==0.1.1
marisa-trie==1.2.1
Markdown==3.6
markdown-it-py==3.0.0
markdown2==2.4.13
MarkupSafe==2.1.5
marshmallow==3.21.1
matplotlib==3.9.2
matplotlib-inline==0.1.6
md2pdf==1.0.1
mdurl==0.1.2
mistune==3.0.2
mmh3==5.1.0
monotonic==1.6
mosestokenizer==1.2.1
mpmath==1.3.0
msgpack==1.0.8
multidict==6.0.5
murmurhash==1.0.10
mypy-extensions==1.0.0
nest-asyncio==1.5.8
newspaper3k==0.2.8
nltk==3.8.1
numpy==1.26.4
oauthlib==3.2.2
onnxruntime==1.21.0
openai==1.68.2
opencv-python==4.9.0.80
openfile==0.0.7
opentelemetry-api==1.27.0
opentelemetry-exporter-otlp-proto-common==1.27.0
opentelemetry-exporter-otlp-proto-grpc==1.27.0
opentelemetry-instrumentation==0.48b0
opentelemetry-instrumentation-asgi==0.48b0
opentelemetry-instrumentation-fastapi==0.48b0
opentelemetry-proto==1.27.0
opentelemetry-sdk==1.27.0
opentelemetry-semantic-conventions==0.48b0
opentelemetry-util-http==0.48b0
ordered-set==4.1.0
orjson==3.10.3
outcome==1.3.0.post0
overrides==7.7.0
packaging==23.2
pandas==2.2.1
parso==0.8.3
pillow==10.2.0
platformdirs==4.1.0
playwright==1.43.0
plotly==5.24.0
pluggy==1.5.0
postgrest==0.16.4
posthog==3.21.0
praw==7.8.1
prawcore==2.4.0
preshed==3.0.9
prompt-toolkit==3.0.43
proto-plus==1.23.0
protobuf==4.25.3
psutil==5.9.7
pure-eval==0.2.2
pyarrow==16.0.0
pyasn1==0.6.0
pyasn1_modules==0.4.0
pycparser==2.22
pydantic==2.7.1
pydantic_core==2.18.2
pydeck==0.9.0
pydyf==0.10.0
pyee==11.1.0
Pygments==2.17.2
PyJWT==2.8.0
PyMuPDF==1.24.2
PyMuPDFb==1.24.1
pyparsing==3.1.2
pypdf==4.2.0
pyphen==0.15.0
PyPika==0.48.9
pypiwin32==223
pyproject_hooks==1.2.0
pyreadline3==3.5.4
PySocks==1.7.1
pytest==8.2.0
pytest-asyncio==0.23.7
pytest-base-url==2.1.0
pytest-playwright==0.4.4
python-dateutil==2.8.2
python-docx==1.1.2
python-dotenv==1.0.1
python-iso639==2024.4.27
python-magic==0.4.27
python-multipart==0.0.9
python-pptx==0.6.21
python-slugify==8.0.4
pytz==2024.1
pywin32==306
PyYAML==6.0.1
pyzmq==25.1.2
rapidfuzz==3.9.1
realtime==1.0.4
referencing==0.35.1
regex==2024.4.28
requests==2.31.0
requests-file==2.1.0
requests-oauthlib==2.0.0
rich==13.7.1
rpds-py==0.18.0
rsa==4.9
sacremoses==0.1.1
safetensors==0.4.3
scikit-learn==1.5.1
scipy==1.12.0
selenium==4.21.0
setuptools==69.1.1
sgmllib3k==1.0.0
shellingham==1.5.4
six==1.16.0
smart-open==7.0.5
smmap==5.0.1
sniffio==1.3.1
sortedcontainers==2.4.0
soupsieve==2.5
sourcedefender==11.0.13
spacy==3.8.2
spacy-legacy==3.0.12
spacy-loggers==1.0.5
SQLAlchemy==2.0.30
srsly==2.4.8
stack-data==0.6.3
starlette==0.37.2
storage3==0.7.4
streamlit==1.34.0
streamlit-authenticator==0.3.2
StrEnum==0.4.15
supabase==2.4.5
supafunc==0.4.5
sympy==1.13.3
tabulate==0.9.0
tavily-python==0.3.3
tenacity==8.2.3
text-unidecode==1.3
TgCrypto==1.2.5
thinc==8.3.2
threadpoolctl==3.5.0
tiktoken==0.7.0
tinycss2==1.3.0
tinysegmenter==0.3
tldextract==5.1.2
tokenizers==0.19.1
toml==0.10.2
toolwrapper==2.1.0
toolz==0.12.1
tornado==6.4
tqdm==4.66.2
traitlets==5.14.1
transformers==4.41.1
trio==0.25.1
trio-websocket==0.11.1
typer==0.12.3
typing-inspect==0.9.0
typing_extensions==4.11.0
tzdata==2024.1
uctools==1.3.0
ujson==5.10.0
unstructured==0.14.2
unstructured-client==0.22.0
update-checker==0.18.0
uritemplate==4.1.1
urllib3==2.2.1
uvicorn==0.29.0
wasabi==1.1.3
watchdog==4.0.0
watchfiles==0.21.0
wcwidth==0.2.13
weasel==0.4.1
weasyprint==62.1
webencodings==0.5.1
websocket-client==1.8.0
websockets==12.0
wheel==0.42.0
wrapt==1.16.0
wsproto==1.2.0
XlsxWriter==3.2.0
yarl==1.9.4
zipp==3.21.0
zopfli==0.2.3
